Update to v1.2.9 + more...
Updated the mod to work on v1.2.9 (probably backwards compatible).

Also removed Net Framework support in favor of .Net Standard 2.0 to remove the hard dependencies of the TaleWorlds DLLs, and also add potential future support for MCM, etc.

Also added an export folder similar to the one used in DCC; however feel free to remove that from this PR if you do not like it.

:thumbsup: